Add stack-threshold functionality to regular unitframe auras.
A blacklist filter will only block an aura if the amount of stacks on the aura is above the set threshold.
A whitelist filter will only allow an aura if the amount of stacks on the aura is above the set threshold.

Added code in core.lua which adds the stackThreshold value to auras if it is missing. This will be true for user-added auras.